+++ angular/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @bnqkl/web-component-angular
 
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @bnqkl/web-component@1.2.1
+
 ## 1.0.5
 
 ### Patch Changes