+++ angular/projects/lib/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @bnqkl/web-component-angular
 
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @bnqkl/web-component@1.2.1
+
 ## 2.0.1
 
 ### Patch Changes